+++ CHANGES.md
@@ -2,8 +2,9 @@
 ## CHANGES [Started from feb 2026]
 
 
-[1.15.x]
- - Added `npm install @anthropic-ai/sdk` like new ai tool provider
+[1.15.xx]
+ - Added `npm install @google/generative-ai` like new ai tool provider.
+ - Added `npm install @anthropic-ai/sdk` like new ai tool provider.
  - Graph part synced - improvment for ai tool - gen graphs
  - Jamb (yacht) game synced with last ver of core + mobile supported.
  - JOLT worker improved for OnContactRemoved and OnContactPersisted (detect) override fn's.